refactor(ppvm-lindblad): address PR #98 review comments
Maintainer-requested code-quality changes (PR #98):

  * PcStepConfig (new config.rs): bundle tau_add / drop_tol / num_threads
    into a config struct passed to pc_step / pc_step_timed (+ inner
    helpers), replacing the long parameter lists and the
    `#[allow(clippy::too_many_arguments)]`. The Python-facing kwargs are
    unchanged — the binding builds the config internally.
  * error.rs (new): move the `Error` enum + Display/Error impls out of
    lib.rs; re-exported as `ppvm_lindblad::Error` (public path unchanged).
  * leakage: merge via Rayon fold/reduce into thread-local maps instead of
    collecting a Vec<HashMap> and merging sequentially — lower peak memory.
    Keys and summed values are identical (FP reduction order differs at
    ~1e-16, well within test tolerances).
  * docs: explain the fixed-width `[u64; W_U64]` Word storage and
    MAX_QUBITS = 64 * W_U64.
  * demo: simplify the redundant position expression to `np.arange(L) - L//2`.

Validated: cargo test --workspace --exclude ppvm-python-native (612 passed),
cargo clippy clean (no too_many_arguments on the refactored fns), and the
full Python suite (136 passed). Python public API unchanged.
